Hash page and refactored pages
