+++ multisig-wallet/ui/app/src/App.tsx
@@ -1,14 +1,9 @@
 import { ThemeProvider } from "@fuel-ui/react";
 import { QueryClientProvider } from "@tanstack/react-query";
 import { queryClient } from "./system/core/utils/queryClient";
-import { BrowserRouter, Route, Routes } from "react-router-dom";
+import { BrowserRouter } from "react-router-dom";
 import { Layout } from "./system/core/components";
-import { ConstructorPage } from "./system/constructor/pages";
-import { ExecuteTransactionPage } from "./system/execute_transaction/pages";
-import { TransferPage } from "./system/transfer/pages";
-import { ThresholdPage } from "./system/threshold/pages";
-import { WeightPage } from "./system/weight/pages";
-import { ViewPage } from "./system/view/pages";
+import { AppRoutes } from "./system/core/components/routes";
 
 function App() {
   return (
@@ -16,14 +11,7 @@ function App() {
       <QueryClientProvider client={queryClient}>
         <ThemeProvider>
           <Layout>
-            <Routes>
-              <Route path="/create" element={<ConstructorPage />} />
-              <Route path="/execute" element={<ExecuteTransactionPage />} />
-              <Route path="/transfer" element={<TransferPage />} />
-              <Route path="/update-threshold" element={<ThresholdPage />} />
-              <Route path="/update-weight" element={<WeightPage />} />
-              <Route path="/view" element={<ViewPage />} />
-            </Routes>
+            <AppRoutes />
           </Layout>
         </ThemeProvider>
       </QueryClientProvider>
