+++ multisig-wallet/ui/app/src/system/hash/pages/index.tsx
@@ -1,0 +1,1 @@
+export * from "./hash";
