+++ multisig-wallet/ui/app/src/system/view/pages/view.tsx
@@ -1,4 +1,4 @@
-import { BoxCentered, Button, Heading, Input, Stack, toast } from "@fuel-ui/react";
+import { BoxCentered, Button, Flex, Heading, Input, Stack, toast } from "@fuel-ui/react";
 import { useContract } from "../../core/hooks";
 import { ContractIdInput } from "../../../contracts/MultisigContractAbi";
 
@@ -18,26 +18,100 @@ export function ViewPage() {
         toast.success(`Balance: ${value}`, { duration: 10000 });
     }
 
+    async function getNonce() {
+        const { value } = await contract!.functions.nonce().get();
+        toast.success(`Current nonce: ${value}`, { duration: 10000 });
+    }
+
+    async function getWeight() {
+        const user = document.querySelector<HTMLInputElement>(
+            `[name="user-weight"]`
+        )!.value;
+
+        const { value } = await contract!.functions.approval_weight(user).get();
+        toast.success(`User weight: ${value}`, { duration: 10000 });
+    }
+
+    async function getThreshold() {
+        const { value } = await contract!.functions.threshold().get();
+        toast.success(`Threshold: ${value}`, { duration: 10000 });
+    }
+
     return (
-        <BoxCentered css={{ marginTop: "8%", width: "30%" }}>
+        <BoxCentered css={{ marginTop: "12%", width: "30%" }}>
 
-            <Stack gap="$2" css={{ minWidth: "100%" }}>
+            <Stack css={{ minWidth: "100%" }}>
 
-                <Heading as="h3" css={{ marginLeft: "auto", marginRight: "auto", color: "$accent1" }}>
-                    Check balance of asset
-                </Heading>
+                <Stack>
+                    <Heading as="h3" css={{ marginLeft: "auto", marginRight: "auto", color: "$accent1" }}>
+                        Check user approval weight
+                    </Heading>
 
-                <Input size="lg">
-                    <Input.Field name="view-asset" placeholder="Asset Id" />
-                </Input>
-                <Button
-                    color="accent"
-                    onPress={getBalance}
-                    size="lg"
-                    variant="solid"
-                >
-                    Get balance
-                </Button>
+                    <Input size="lg">
+                        <Input.Field name="user-weight" placeholder="User address" />
+                    </Input>
+
+                    <Button
+                        color="accent"
+                        onPress={getWeight}
+                        size="lg"
+                        variant="solid"
+                    >
+                        Get weight
+                    </Button>
+                </Stack>
+
+                <Stack css={{ minWidth: "100%", marginTop: "$10" }}>
+
+                    <Heading as="h3" css={{ marginLeft: "auto", marginRight: "auto", color: "$accent1" }}>
+                        Check balance of asset
+                    </Heading>
+
+                    <Input size="lg">
+                        <Input.Field name="view-asset" placeholder="Asset Id" />
+                    </Input>
+                    <Button
+                        color="accent"
+                        onPress={getBalance}
+                        size="lg"
+                        variant="solid"
+                    >
+                        Get balance
+                    </Button>
+                </Stack>
+
+                <Flex gap="$1" css={{ minWidth: "100%", marginTop: "$10" }}>
+                    <Stack css={{ minWidth: "50%" }}>
+                        <Heading as="h3" css={{ marginLeft: "auto", marginRight: "auto", marginTop: "$14", color: "$accent1" }}>
+                            Check nonce
+                        </Heading>
+
+                        <Button
+                            color="accent"
+                            onPress={getNonce}
+                            size="lg"
+                            variant="solid"
+                        >
+                            Get nonce
+                        </Button>
+                    </Stack>
+
+                    <Stack css={{ minWidth: "50%" }}>
+                        <Heading as="h3" css={{ marginLeft: "auto", marginRight: "auto", marginTop: "$14", color: "$accent1" }}>
+                            Check threshold
+                        </Heading>
+
+                        <Button
+                            color="accent"
+                            onPress={getThreshold}
+                            size="lg"
+                            variant="solid"
+                        >
+                            Get threshold
+                        </Button>
+                    </Stack>
+                </Flex>
+
             </Stack>
             
         </BoxCentered>
